:recycle: Refactor: validation 위치 도메인생성자로 변경
