+++ src/main/java/com/waterfogsw/voucher/voucher/service/VoucherManageService.java
@@ -11,20 +11,6 @@ public class VoucherManageService implements VoucherService {
 
     @Override
     public Voucher saveVoucher(Voucher voucher) {
-        validateVoucher(voucher);
         return voucherRepository.save(voucher);
-    }
-
-    private void validateVoucher(Voucher voucher) {
-        switch (voucher.getType()) {
-            case FIXED_AMOUNT -> {
-                if(voucher.getValue() < 0)
-                    throw new IllegalArgumentException();
-            }
-            case PERCENT_DISCOUNT -> {
-                if(voucher.getValue() < 0 || voucher.getValue() > 100)
-                    throw new IllegalArgumentException();
-            }
-        }
     }
 }
