+++ src/test/java/com/waterfogsw/voucher/voucher/VoucherServiceTests.java
@@ -35,33 +35,6 @@ public class VoucherServiceTests {
     class Describe_addVoucher {
 
         @Nested
-        @DisplayName("type 이 FIXED_AMOUNT 인 Voucher 의 value 가 음수이면")
-        class Context_with_negative_fixedAmount {
-
-            @Test
-            @DisplayName("IllegalArgumentException 예외를 발생시킨다.")
-            void it_throw_IllegalArgumentException() {
-                Voucher voucher = new Voucher(VoucherType.FIXED_AMOUNT, -1000);
-
-                assertThrows(IllegalArgumentException.class, () -> voucherService.saveVoucher(voucher));
-            }
-        }
-
-        @Nested
-        @DisplayName("type 이 PERCENT_DISCOUNT 인 Voucher 의 value 가 1 ~ 100 사이의 수가 아니면")
-        class Context_with_out_of_range_percent_discount {
-
-            @ParameterizedTest
-            @ValueSource(ints = {-1, 101})
-            @DisplayName("IllegalArgumentException 예외를 발생시킨다.")
-            void it_throw_IllegalArgumentException(int percent) {
-                Voucher voucher = new Voucher(VoucherType.PERCENT_DISCOUNT, percent);
-
-                assertThrows(IllegalArgumentException.class, () -> voucherService.saveVoucher(voucher));
-            }
-        }
-
-        @Nested
         @DisplayName("Voucher 가 정상적으로 저장되어 생성되면")
         class Context_with_saved_success {
